First sketch of a private logging and redo infrastructure
- plog mechanism still very rudimentary; enventually we'll have a
thread_local transaction context object that will manage private logs
- Log record generation and other operations like copying, sorting,
moving, etc, only work with a log record header and a payload area, on
which serialization works in an abstract way, using the VariadicEncoder.
Payload only needs to be interpreted if we need to redo, inspect, or
print the log record.
- Only redo of insert log record implemented so far, but infrastructure
is there